Crate expected

An assertion utility focused on unit testing.

Macros

expect
Declare an expectation that the specified conditional expression is true.

expect_eq
Declare an expectation that the specified values are equal.

Structs

Disappoint

Disappoints
A set of disappoints occurred during an execution of asseverate.

Expected

Traits

FutureExpectedExt

Functions

expected
Run the provided closure and checks to see if all expectation have been satisfied.